+++ wikimedia/json-codec/src/CodecService.php
@@ -1,0 +1,43 @@
+<?php
+declare( strict_types=1 );
+
+/**
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * http://www.gnu.org/copyleft/gpl.html
+ *
+ * @file
+ */
+
+namespace Wikimedia\JsonCodec;
+
+use Attribute;
+
+/**
+ * Names the service container entry that should be supplied for the
+ * annotated parameter of ::toJsonArray() or ::newFromJsonArray().
+ *
+ * Optional services can set `optional` to `true`, in which case
+ * `null` is passed if the service is not available.
+ *
+ * @see JsonCodecableWithServicesTrait
+ */
+#[Attribute( Attribute::TARGET_PARAMETER )]
+class CodecService {
+	public function __construct(
+		public readonly string $name,
+		public readonly bool $optional = false,
+	) {
+	}
+}